services/flask/requirements.txt
Flask==1.1.2
gunicorn==20.0.4
requests==2.20.1
flask-mongoengine==1.0.0
flask-graphql==2.0.1
graphene-mongo==0.2.13
promise==2.3
